requirements.txt
streamlit==1.32.0
openai==1.12.0
groq==0.4.2
httpx==0.25.2
httpcore==1.0.2
python-dotenv==1.0.1
requests==2.31.0
pandas==2.0.3
plotly==5.19.0
numpy==1.24.3
scikit-learn==1.3.2
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
nltk>=3.8.1
beautifulsoup4>=4.12.0
lxml>=4.9.0
numpy>=1.24.0